+++ runtime/crab/src/pallets/mod.rs
@@ -1,0 +1,8 @@
+pub mod system;
+pub use system::*;
+
+pub mod babe;
+pub use babe::*;
+
+pub mod timestamp;
+pub use timestamp::*;
